+++ src/rest/api/v2010/account/message.ts
@@ -73,13 +73,13 @@ export interface MessageListInstanceCreateOptions {
   statusCallback?: string;
   /** The SID of the associated [TwiML Application](https://www.twilio.com/docs/usage/api/applications). [Message status callback requests](https://www.twilio.com/docs/sms/api/message-resource#twilios-request-to-the-statuscallback-url) are sent to the TwiML App\\\'s `message_status_callback` URL. Note that the `status_callback` parameter of a request takes priority over the `application_sid` parameter; if both are included `application_sid` is ignored. */
   applicationSid?: string;
-  /** [DEPRECATED] This parameter will no longer have any effect as of 2024-06-03. */
+  /** [OBSOLETE] This parameter will no longer have any effect as of 2024-06-03. */
   maxPrice?: number;
   /** Boolean indicating whether or not you intend to provide delivery confirmation feedback to Twilio (used in conjunction with the [Message Feedback subresource](https://www.twilio.com/docs/sms/api/message-feedback-resource)). Default value is `false`. */
   provideFeedback?: boolean;
   /** Total number of attempts made (including this request) to send the message regardless of the provider used */
   attempt?: number;
-  /** The maximum length in seconds that the Message can remain in Twilio\\\'s outgoing message queue. If a queued Message exceeds the `validity_period`, the Message is not sent. Accepted values are integers from `1` to `14400`. Default value is `14400`. A `validity_period` greater than `5` is recommended. [Learn more about the validity period](https://www.twilio.com/blog/take-more-control-of-outbound-messages-using-validity-period-html) */
+  /** The maximum length in seconds that the Message can remain in Twilio\\\'s outgoing message queue. If a queued Message exceeds the `validity_period`, the Message is not sent. Accepted values are integers from `1` to `36000`. Default value is `36000`. A `validity_period` greater than `5` is recommended. [Learn more about the validity period](https://www.twilio.com/blog/take-more-control-of-outbound-messages-using-validity-period-html) */
   validityPeriod?: number;
   /** Reserved */
   forceDelivery?: boolean;